+++ R/gdalraster.R
@@ -63,6 +63,7 @@
 #'
 #' ds$getMetadata(band, domain)
 #' ds$getMetadataItem(band, mdi_name, domain)
+#' ds$setMetadataItem(band, mdi_name, mdi_value, domain)
 #'
 #' ds$read(band, xoff, yoff, xsize, ysize, out_xsize, out_ysize)
 #' ds$write(band, xoff, yoff, xsize, ysize, rasterData)
@@ -314,6 +315,13 @@
 #' band number to retrieve band-level metadata.
 #' Set \code{domain = ""} (empty string) to retrieve an item in the 
 #' default domain.
+#'
+#' \code{$setMetadataItem(band, mdi_name, mdi_value, domain)}
+#' Sets the value (\code{mdi_value}) of a specific metadata item named
+#' \code{mdi_name} in the specified \code{domain}.
+#' Set \code{band = 0} to set dataset-level metadata, or to an integer 
+#' band number to set band-level metadata.
+#' Set \code{domain = ""} (empty string) to set an item in the default domain.
 #'
 #' \code{$read(band, xoff, yoff, xsize, ysize, out_xsize, out_ysize)}
 #' Reads a region of raster data from \code{band}. The method takes care of